LStorage: Added a footer, restructured
