+++ src/js/components/footer.js
@@ -1,0 +1,3 @@
+const footer = document.createElement("footer");
+// footer.innerHTML = style
+document.documentElement.appendChild(footer);
